Fix improper use of 198.51.0.0/24
The above prefix is just a regular global prefixed used by someone out
there on the internet. It was originally meant to be 198.51.0.0/24
(TEST-NET-2 from RFC 5737), however instead replace it with
192.0.2.0/24 (TEST-NET-1) in order to prevent the figures from
becoming too wide.

Change the single pre-existing use of 192.0.2.0/24 in siit-dc section
3.2 to 198.51.0.0/24 to avoid using the same prefix for different
purposes.

Thanks for pointing out this mistake, Fred Baker!